stackapi: Move stackapi up one level
