+++ pkg/api/stackapi/delete.go
