+++ pkg/api/stackapi/delete_test.go
